Remove workflow.go file and move func to helpers.go
