+++ pkg/sdk/renderer/argo/helpers.go
@@ -29,3 +29,17 @@ func shouldExit(ctx context.Context) bool {
 		return false
 	}
 }
+
+func getEntrypointWorkflowIndex(w *Workflow) (int, error) {
+	if w == nil {
+		return 0, NewWorkflowNilError()
+	}
+
+	for idx, tmpl := range w.Templates {
+		if tmpl.Name == w.Entrypoint {
+			return idx, nil
+		}
+	}
+
+	return 0, NewEntrypointWorkflowIndexNotFoundError(w.Entrypoint)
+}
